Rename app to Api, move api into package
